pyproject.toml
[tool.poetry]
name = "investment-app"
version = "0.1.0"
description = ""
authors = ["Jakub Turner <kuba.turner@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
mysql-connector-python = "^8.0.30"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"